requirements.txt
fastapi==0.100.0
uvicorn==0.23.1
google-cloud-aiplatform==1.67.1
google-auth==2.35.0
google-api-core==2.20.0
google-generativeai==0.8.1
google-api-python-client==2.146.0
attrs==24.2.0
certifi==2024.8.30
charset-normalizer==3.3.2
idna==3.10
requests==2.32.3
